보정: CMYK JPEG 계약을 integration suite로 이동
